Fresh Start Master

+++ db/migrate/02_create_tweets.rb
@@ -7,7 +7,7 @@ class CreateTweets < ActiveRecord::Migration[5.1]
       t.integer :favorite_count
       t.integer :retweet_count
       t.string :lang
-      t.integer :tweet_code
+      t.bigint :tweet_code
     end
   end
 end
